4 teensy bugs fixed: 1) body margin that has been haunting me, 2) allow unique usernames only, 3) confirm message prior to deletion, 4) inventory card description overflow

+++ client/src/Components/Details.jsx
@@ -46,23 +46,23 @@ const Details = () => {
   const handleDelete = () => {
     const requestID = details.id;
 
-    fetch(`http://localhost:8080/items/${requestID}`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json'
-      }
-    })
-    .then(res => res.json())
-    .then(data => {
-      if(data.message) {
-        alert(data.message)
-      }
-    })
-    .then(() => {navigate('/inventory')})
-    .catch(err => console.log(err))
+    if (confirm(`Select 'OK' to delete ${details.item_name}.`)) {
+      fetch(`http://localhost:8080/items/${requestID}`, {
+        method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json'
+        }
+      })
+      .then(res => res.json())
+      .then(data => {
+        if(data.message) {
+          alert(data.message)
+        }
+      })
+      .then(() => {navigate('/inventory')})
+      .catch(err => console.log(err))
+    }
   }
-
-  console.log(newDetails)
 
   return(
     <>
